Add runtime option to use a classname instead of inline styles
