+++ .github/lock.yml
@@ -2,17 +2,7 @@
 
 # Number of days of inactivity before a closed issue or pull request is locked
 daysUntilLock: 5
+
+# Do not comment when locking
 setLockReason: false
-
-issues:
-  # Comment to post when locking a stale Issue.
-  lockComment: >
-    The conversation for this issue has been locked because the issue is
-    closed. If you have a question or comment, please open a new issue.
-
-pulls:
-# Comment to post when locking a stale Pull Request.
-  lockComment: >
-    The conversation for this pull request has been locked because the pull
-    request is closed. If you would like to contribute please open a new pull
-    request.
+lockComment: false
