+++ .travis.yml
@@ -4,6 +4,7 @@ language: node_js
 node_js:
   - 8
   - 10
+  - node
 cache:
   yarn: true
   directories:
