+++ docusaurus/docs/developing-components-in-isolation.md
@@ -22,16 +22,10 @@ You can also deploy your Storybook or style guide as a static app. This way, eve
 
 Storybook is a development environment for React UI components. It allows you to browse a component library, view the different states of each component, and interactively develop and test components.
 
-First, install the following npm package globally:
+Run the following command inside your app’s directory:
 
 ```sh
-npm install -g @storybook/cli
-```
-
-Then, run the following command inside your app’s directory:
-
-```sh
-getstorybook
+npx -p @storybook/cli sb init
 ```
 
 After that, follow the instructions on the screen.
