+++ packages/babel-plugin-named-asset-import/index.test.js
@@ -46,5 +46,53 @@ pluginTester({
         'import { logoUrl } from "logo.svg";\n' +
         'import { ReactComponent as Logo } from "@svgr/webpack?-prettier,-svgo!logo.svg";',
     },
+    defaultExport: {
+      code: 'export default logo;',
+      output: 'export default logo;',
+    },
+    constExport: {
+      code: 'export const token = "token";',
+      output: 'export const token = "token";',
+    },
+    classExport: {
+      code: 'export class Logo {}',
+      output: 'export class Logo {}',
+    },
+    namedExport: {
+      code: 'export { logo } from "logo";',
+      output: 'export { logo } from "logo";',
+    },
+    namedExportRenamed: {
+      code: 'export { Url as logo } from "logo";',
+      output: 'export { Url as logo } from "logo";',
+    },
+    allExport: {
+      code: 'export * from "logo";',
+      output: 'export * from "logo";',
+    },
+    svgNamedExport: {
+      code: 'export { logo } from "logo.svg";',
+      output: 'export { logo } from "logo.svg";',
+    },
+    svgAllExport: {
+      code: 'export * from "logo.svg";',
+      output: 'export * from "logo.svg";',
+    },
+    svgReactComponentNamedExport: {
+      code: 'export { ReactComponent as Logo } from "logo.svg";',
+      output:
+        'export { ReactComponent as Logo } from "@svgr/webpack?-prettier,-svgo!logo.svg";',
+    },
+    svgReactComponentExport: {
+      code: 'export { ReactComponent } from "logo.svg";',
+      output:
+        'export { ReactComponent } from "@svgr/webpack?-prettier,-svgo!logo.svg";',
+    },
+    svgMultipleExport: {
+      code: 'export { logoUrl , ReactComponent as Logo } from "logo.svg";',
+      output:
+        'export { logoUrl } from "logo.svg";\n' +
+        'export { ReactComponent as Logo } from "@svgr/webpack?-prettier,-svgo!logo.svg";',
+    },
   },
 });
