+++ packages/babel-preset-react-app/package.json
@@ -17,24 +17,24 @@
     "test.js"
   ],
   "dependencies": {
-    "@babel/core": "7.1.0",
+    "@babel/core": "7.1.6",
     "@babel/plugin-proposal-class-properties": "7.1.0",
-    "@babel/plugin-proposal-decorators": "7.1.2",
+    "@babel/plugin-proposal-decorators": "7.1.6",
     "@babel/plugin-proposal-object-rest-spread": "7.0.0",
     "@babel/plugin-syntax-dynamic-import": "7.0.0",
     "@babel/plugin-transform-classes": "7.1.0",
-    "@babel/plugin-transform-destructuring": "7.0.0",
-    "@babel/plugin-transform-flow-strip-types": "7.0.0",
+    "@babel/plugin-transform-destructuring": "7.1.3",
+    "@babel/plugin-transform-flow-strip-types": "7.1.6",
     "@babel/plugin-transform-react-constant-elements": "7.0.0",
     "@babel/plugin-transform-react-display-name": "7.0.0",
     "@babel/plugin-transform-runtime": "7.1.0",
-    "@babel/preset-env": "7.1.0",
+    "@babel/preset-env": "7.1.6",
     "@babel/preset-react": "7.0.0",
     "@babel/preset-typescript": "7.1.0",
-    "@babel/runtime": "7.0.0",
+    "@babel/runtime": "7.1.5",
     "babel-loader": "8.0.4",
     "babel-plugin-dynamic-import-node": "2.2.0",
     "babel-plugin-macros": "2.4.2",
-    "babel-plugin-transform-react-remove-prop-types": "0.4.18"
+    "babel-plugin-transform-react-remove-prop-types": "0.4.20"
   }
 }
