+++ packages/react-app-polyfill/ie11.js
@@ -14,8 +14,12 @@ if (typeof Promise === 'undefined') {
   window.Promise = require('promise/lib/es6-extensions.js');
 }
 
-// fetch() polyfill for making API calls.
-require('whatwg-fetch');
+// Make sure we're in a Browser-like environment before importing polyfills
+// This prevents `fetch()` from being imported in a Node test environment
+if (typeof window !== 'undefined') {
+  // fetch() polyfill for making API calls.
+  require('whatwg-fetch');
+}
 
 // Object.assign() is commonly used with React.
 // It will use the native implementation if it's present and isn't buggy.
