+++ test/fixtures/__shared__/template/public/index.html
@@ -1,5 +1,5 @@
 <!DOCTYPE html>
-  <html lang="en">
+<html lang="en">
   <head>
     <title>React App</title>
   </head>
