+++ test/fixtures/typescript/src/App.test.ts
@@ -6,3 +6,10 @@ it('reads a typescript file with no syntax error', () => {
   expect(App.foo.baz!.n).toBe(123);
   expect(app.n).toBe(123);
 });
+
+it('supports decorators', () => {
+  expect((App as any).annotated).toBe(true);
+
+  const app = new App();
+  expect(app.decorated).toBe(42);
+});
